requirements.txt
asgiref==3.2.3
Django==3.0.1
pytz==2019.3
sqlparse==0.3.0
django-taggit==1.2.0
psycopg2==2.8.4
